requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-settings==2.0.3
uvicorn[standard]==0.24.0

httpx<0.28
openai==1.3.7
numpy==1.24.3
tiktoken==0.5.1

asyncpg==0.29.0
pgvector==0.2.4
psycopg2-binary==2.9.7
